refactor: move FavoriteControllerXX to profile domain
